refactor: move shared resource-policy context into core
runner (core) read the process-captured resource-policy context and the
runner-placement environment probes from commands::utils::resource_policy,
creating a core -> commands dependency edge that blocks extracting commands
into its own crate.

Move the shared pieces down into a new core::resource_policy_context module:
the ResourcePolicyContext data type (plus RunnerSelection/HostSnapshot
sub-structs), the process-wide capture store (capture_context/captured_context),
and the runner-placement env probes (is_managed_runner_placement_context et al,
which already read core::runner env constants). The CLI-specific evaluation
logic that builds a context from a DoctorOutput stays in commands.

commands::utils::resource_policy re-exports the moved items so its existing call
sites are unchanged; runner now reads them directly from core. This removes two
of the three real core -> commands edges (the remaining one, an AgentTaskCommand
CLI-parse in runner, is a larger follow-up tied to cli_surface layering).

cargo check --lib and --lib --tests pass.
