+++ src/core/mod.rs
@@ -143,6 +143,7 @@ pub mod report_compare;
 pub(crate) mod report_compare_render;
 pub mod resource_cleanup_intent;
 pub mod resource_lifecycle_index;
+pub mod resource_policy_context;
 pub mod resources;
 pub mod review;
 pub mod rig;
